# Improve the performance of client-server array transfers