--- a/benchmarks/graph_infra/array_transfer.perfkeys
+++ b/benchmarks/graph_infra/array_transfer.perfkeys
@@ -1,0 +1,4 @@
+to_ndarray Average time =
+to_ndarray Average rate =
+ak.array Average time =
+ak.array Average rate =