--- a/benchmarks/run_benchmarks.py
+++ b/benchmarks/run_benchmarks.py
@@ -20,8 +20,9 @@ from util import *
 logging.basicConfig(level=logging.INFO)
 
 BENCHMARKS = ['stream', 'argsort', 'coargsort', 'groupby', 'aggregate', 'gather', 'scatter',
-              'reduce', 'scan', 'noop', 'setops', 'array_create', 'IO',
-              'str-argsort', 'str-coargsort', 'str-groupby', 'str-gather']
+              'reduce', 'scan', 'noop', 'setops', 'array_create',
+              'array_transfer', 'IO', 'str-argsort', 'str-coargsort',
+              'str-groupby', 'str-gather']
 
 def get_chpl_util_dir():
     """ Get the Chapel directory that contains graph generation utilities. """